+++ extended_mypy_django_plugin/plugin/actions/_sem_analyze.py
@@ -1,17 +1,24 @@
 from mypy.nodes import (
     GDEF,
+    CastExpr,
     StrExpr,
     SymbolTableNode,
     TypeInfo,
     TypeVarExpr,
+    Var,
 )
 from mypy.plugin import AnalyzeTypeContext, DynamicClassDefContext
 from mypy.semanal import SemanticAnalyzer
 from mypy.typeanal import TypeAnalyser
 from mypy.types import (
     AnyType,
+    CallableType,
+    Instance,
     TypeOfAny,
+    TypeType,
     TypeVarType,
+    UnionType,
+    get_proper_type,
 )
 from mypy.types import (
     Type as MypyType,
@@ -74,6 +81,98 @@ class SemAnalyzing:
             return instance.type
 
         return self.store.plugin_lookup_info(fullname)
+
+    def transform_cast_as_concrete(self, ctx: DynamicClassDefContext) -> None:
+        if len(ctx.call.args) != 1:
+            self.api.fail("Concrete.cast_as_concrete takes exactly one argument", ctx.call)
+            return None
+
+        node = self.api.lookup_type_node(ctx.call.args[0])
+        if not node or not node.node:
+            self.api.fail("Failed to lookup the argument", ctx.call)
+            return None
+
+        arg_node = self.api.lookup_current_scope(node.node.name)
+
+        if arg_node is None or arg_node.type is None or arg_node.node is None:
+            return None
+
+        if not isinstance(arg_node.node, Var):
+            return None
+
+        arg_node_typ = get_proper_type(arg_node.type)
+
+        is_type: bool = False
+        if isinstance(arg_node_typ, TypeType):
+            is_type = True
+            arg_node_typ = arg_node_typ.item
+
+        if not isinstance(arg_node_typ, Instance | UnionType | TypeVarType):
+            self.api.fail(
+                f"Unsure what to do with the type of the argument given to cast_as_concrete: {arg_node_typ}",
+                ctx.call,
+            )
+            return None
+
+        if isinstance(arg_node_typ, TypeVarType):
+            if (
+                self.api.is_func_scope()
+                and isinstance(self.api.type, TypeInfo)
+                and arg_node_typ.name == "Self"
+            ):
+                replacement: Instance | TypeType | None = self.api.named_type_or_none(
+                    self.api.type.fullname
+                )
+                if replacement:
+                    arg_node_typ = replacement
+                    if is_type:
+                        replacement = TypeType(replacement)
+
+                    if self.api.scope.function and isinstance(
+                        self.api.scope.function.type, CallableType
+                    ):
+                        if self.api.scope.function.type.arg_names[0] == ctx.name:
+                            # Avoid getting an assignment error trying to assign a union of the concrete types to
+                            # a variable typed in terms of Self
+                            self.api.scope.function.type.arg_types[0] = replacement
+                else:
+                    self.api.fail("Failed to resolve Self", ctx.call)
+                    return None
+            else:
+                self.api.fail(
+                    f"Resolving type variables for cast_as_concrete not implemented: {arg_node_typ}",
+                    ctx.call,
+                )
+                return None
+
+        def defer() -> bool:
+            if self.api.final_iteration:
+                return True
+            else:
+                self.api.defer()
+                return False
+
+        resolver = _annotation_resolver.AnnotationResolver(
+            self.store,
+            defer=defer,
+            fail=lambda msg: self.api.fail(msg, ctx.call),
+            lookup_info=self._lookup_info,
+            named_type_or_none=self.api.named_type_or_none,
+        )
+
+        concrete = resolver.resolve(
+            _known_annotations.KnownAnnotations.CONCRETE,
+            TypeType(arg_node_typ) if is_type else arg_node_typ,
+        )
+        if not concrete:
+            return None
+
+        ctx.call.analyzed = CastExpr(ctx.call.args[0], concrete)
+        ctx.call.analyzed.line = ctx.call.line
+        ctx.call.analyzed.column = ctx.call.column
+        ctx.call.analyzed.accept(self.api)
+
+        return None
 
     def transform_type_var_classmethod(self, ctx: DynamicClassDefContext) -> None:
         if not isinstance(ctx.call.args[0], StrExpr):